+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added 
+- Minimum read quality filter.
+### Changed
+- In the report 'reads' now refers to number of reads not subreads
+ 
 ## [v2.3.0]
 ### Fix
 - Output schema with correct expression matrix paths.